requirements.txt
# addict
# future
# DISTS-pytorch
einops
# lmdb
lpips
matplotlib
opencv-python-headless
# pandas
# Pillow
python-dotenv
pytorch-lightning
# PyYAML
# requests
scipy
timm==0.9.2
torch==2.0
tensorboard
# tqdm